+++ app/components/MenuAppBar.js
@@ -7,6 +7,8 @@ import { DebounceInput } from 'react-debounce-input';
 import { compose } from 'recompose';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
+import Tooltip from '@material-ui/core/Tooltip';
+import Badge from '@material-ui/core/Badge';
 import Typography from '@material-ui/core/Typography';
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
@@ -21,13 +23,8 @@ import { updateSearchInput, clearSearch } from '../actions/bundleFilter.actions'
 
 
 function mapStateToProps(state, props) {
-  const { bundlesFilter, authentication, bundleManageResources } = state;
-  const { selectedResourcesToPaste = {} } = bundleManageResources;
-  const clipboard = {
-    bundleId: selectedResourcesToPaste.bundleId,
-    description: 'Resources',
-    items: selectedResourcesToPaste.uris || []
-  };
+  const { bundlesFilter, authentication, clipboard: clipboardState } = state;
+  const { selectedItemsToPaste: clipboard = {} } = clipboardState;
   const { isLoading: isLoadingSearch } = bundlesFilter;
   const { isSearchActive } = bundlesFilter;
   const { searchInputRaw } = bundlesFilter;
@@ -59,8 +56,7 @@ type Props = {
     showSearch?: boolean,
     showClipboard?: boolean,
     clipboard: ?{},
-    updateSearchInput: () => {},
-    clearSearch: () => {}
+    updateSearchInput: () => {}
 };
 
 const styles = theme => ({
@@ -75,6 +71,14 @@ const styles = theme => ({
   },
   iconSmall: {
     fontSize: 20,
+  },
+  iconSmaller: {
+    fontSize: 12,
+  },
+  badge: {
+    marginRight: 7,
+    height: 18,
+    width: 18
   },
   root: {
     flexGrow: 1,
@@ -125,7 +129,8 @@ class MenuAppBar extends React.PureComponent {
     } = this.props;
     const { anchorEl } = this.state;
     const open = Boolean(anchorEl);
-
+    const clipboardMedium = clipboard.bundleId ? clipboard.getMedium() : '';
+    const clipboardTooltip = clipboard.bundleId ? `${clipboard.itemsType} from ${clipboard.getDisplayAs().name} ${clipboard.getDisplayAs().revision}` : '';
     return (
       <AppBar position="sticky">
         <Toolbar>
@@ -147,17 +152,21 @@ class MenuAppBar extends React.PureComponent {
           </div>}
           {showClipboard && clipboard.bundleId &&
           <div style={{ marginLeft: '10px', marginRight: '10px' }}>
-            <Button
-              key="btnClipboard"
-              color="inherit"
-              // onClick={this.handlePasteResources}
-            >
-              <AssignmentIcon className={classNames(classes.leftIcon)} />
-              {ux.conditionallyRenderBadge(
-                  { classes: { badge: classes.badgeTight }, color: 'secondary' }, clipboard.items.length,
-                  ''
-                  )}
-            </Button>
+            <Tooltip title={clipboardTooltip}>
+              <Button
+                key="btnClipboard"
+                color="inherit"
+                // onClick={this.handlePasteResources}
+              >
+                <Badge badgeContent={ux.getMediumIcon(clipboardMedium, { className: classNames(classes.rightIcon, classes.iconSmaller) })} >
+                  <AssignmentIcon className={classNames(classes.leftIcon)} />
+                </Badge>
+                {ux.conditionallyRenderBadge(
+                    { classes: { badge: classes.badge }, color: 'secondary' }, clipboard.items.length,
+                    ''
+                    )}
+              </Button>
+            </Tooltip>
           </div>
           }
           {workspaceName && (