+++ app/services/bundle.service.js
@@ -35,15 +35,17 @@ export const bundleService = {
   getFormFields,
   deleteForm,
   postFormFields,
+  waitUntilPostFormFields,
   startUploadBundle,
   startCreateContent,
   stopCreateContent,
   bundleIsInCreateMode,
   waitStartCreateMode,
   waitStopCreateMode,
-  waitMode,
+  waitUntilBundleCondition,
   postResource,
   copyResources,
+  copyMetadata,
   forkBundle,
   updateManifestResource,
   getApplicableWizards,
@@ -354,6 +356,12 @@ function copyResources(bundleId, fromBundleId, uris, merge = true) {
   return bundleAddTasks(bundleId, copyResourcesTask);
 }
 
+function copyMetadata(bundleId, fromBundleId, sections, preserveExisting = true) {
+  const sectionsList = sections.map(section => `<section>${section}</section>`).join('');
+  const copyResourcesTask = `<copyMetadata><fromBundleId>${fromBundleId}</fromBundleId>${sectionsList}<preserveExisting>${preserveExisting}</preserveExisting></copyMetadata>`;
+  return bundleAddTasks(bundleId, copyResourcesTask);
+}
+
 function bundleAddTasks(bundleId, innerTasks) {
   const requestOptions = {
     method: 'POST',
@@ -428,6 +436,13 @@ function getFormFields(bundleId, formKey) {
     "response_valid": false
   }
  */
+async function waitUntilPostFormFields(postFormFieldArgs) {
+  const { bundleId } = postFormFieldArgs;
+  await bundleService.waitStartCreateMode(bundleId);
+  const response = await postFormFields(postFormFieldArgs);
+  return response;
+}
+
 function postFormFields({
   bundleId, formKey, payload, keyField
 }) {
@@ -500,9 +515,9 @@ function getCurrentBundleState(getState, bundleId) {
   return bundleId ? addedByBundleIds[bundleId] : null;
 }
 
-async function waitMode(getState, bundleId, mode) {
+async function waitUntilBundleCondition(getState, bundleId, condition) {
   const bundleCurrentState = bundleService.getCurrentBundleState(getState, bundleId);
-  await waitUntil(async () => bundleCurrentState.mode === mode, 60000, 500);
+  await waitUntil(async () => condition(bundleCurrentState), 60000, 500);
   return bundleCurrentState;
 }